feat: implement file upload functionality; add validation and cloud storage integration

+++ Backend/src/utils/cloudinary.js
@@ -1,5 +1,6 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import ApiError from "./ApiError.js";
 
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
@@ -9,21 +10,17 @@ cloudinary.config({
 
 export const uploadToCloudinary = async (filePath) => {
     try {
-        if (!filePath) return null;
         const response = await cloudinary.uploader.upload(filePath, {
             resource_type: "auto"
         })
 
-        if (fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath);
-        }
+        if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+
         return response;
 
     } catch (error) {
-        if (fs.existsSync(filePath)) {
-            fs.unlinkSync(filePath);
-        }
-        return null;
+        if (fs.existsSync(filePath)) fs.unlinkSync(filePath);
+        throw new ApiError(500, `Cloudinary upload failed: ${error.message}`);
     }
 }
